refactor(core): move CacheableSource into diracx.core.sources
Co-locate the sync CacheableSource base class with its async sibling
AsyncCacheableSource in diracx.core.sources, leaving diracx.core.config.sources
to hold only config-specific sources. CacheableSource is fully generic and only
lived in the config module for historical reasons.

diracx.core.config still re-exports CacheableSource, so existing imports keep
working.
